Admin

+++ app/Controllers/Backend/News.php
@@ -191,7 +191,7 @@ class News extends BaseController
 			if (!$validation->withRequest($this->request)->run()) {
 				$response =	$this->field->errorValidation($this->table);
 			} else {
-				if (!empty($imgName)) {
+				if (!empty($imgName) && file_exists($row->md_image_id)) {
 					// Remove old image path directory
 					unlink($this->path_folder . $row->image);
 
@@ -205,6 +205,16 @@ class News extends BaseController
 						// Move to folder
 						$file->move($this->path_folder, $newfilename);
 					}
+				} else if (!empty($imgName)) {
+					if (!empty($row->image_id)) {
+						// Delete old image data
+						$delete = $image->delete($row->image_id);
+					}
+					// Insert new data image
+					$image_id = $image->insert_image($newfilename, $this->path_folder);
+
+					// Move to folder
+					$file->move($this->path_folder, $newfilename);
 				}
 
 				if (isset($image_id)) {
@@ -228,7 +238,7 @@ class News extends BaseController
 
 		$image_id = 0;
 
-		$row = $news->detail('trx_news_id', $id)->getRow();
+		$row = $principal->detail('trx_news_id', $id)->getRow();
 
 		if (!empty($row->image_id)) {
 			$image_id = $row->image_id;
