+++ app/Controllers/Backend/Principal.php
@@ -178,7 +178,7 @@ class Principal extends BaseController
 			if (!$validation->withRequest($this->request)->run()) {
 				$response = $this->field->errorValidation($this->table);
 			} else {
-				if (!empty($imgName)) {
+				if (!empty($imgName) && file_exists($row->md_image_id)) {
 					// Remove old image path directory
 					unlink($this->path_folder . $row->image);
 
@@ -192,6 +192,17 @@ class Principal extends BaseController
 						// Move to folder
 						$file->move($this->path_folder, $newfilename);
 					}
+				} else if (!empty($imgName)) {
+					if (!empty($row->image_id)) {
+						// Delete old image data
+						$delete = $image->delete($row->image_id);
+					}
+
+					// Insert new data image
+					$image_id = $image->insert_image($newfilename, $this->path_folder);
+
+					// Move to folder
+					$file->move($this->path_folder, $newfilename);
 				}
 
 				if (isset($image_id)) {
